优化了传参，增加了gitignore文件

+++ AIDetector_pytorch.py
@@ -20,7 +20,7 @@ class Detector(baseDet):
         self.device = select_device(self.device)
         model = attempt_load(self.weights, map_location=self.device)
         model.to(self.device).eval()
-        model.half()
+        model.float()
         # torch.save(model, 'test.pt')
         self.m = model
         self.names = model.module.names if hasattr(
@@ -43,8 +43,7 @@ class Detector(baseDet):
     def detect(self, im):
 
         im0, img = self.preprocess(im)
-
-        pred = self.m(img, augment=False)[0]
+        pred = self.m(img.float(), augment=False)[0]
         pred = pred.float()
         pred = non_max_suppression(pred, self.threshold, 0.4)
